+++ README.md
@@ -1,2 +1,12 @@
 # Repository-for-the-IUTLL
 Repository for all kinds of programs created by the students of the IUTLL
+
+all the programs and exercises are separated by programming language for more specific searches.
+Each language has a topic like "data structures" for example
+
+____
+
+# Repositorio para el IUTLL
+Repositorio con toda clase de programas creados por estudiantes del IUTLL.
+
+todos los programas están separados por lenguajes de programación para busquedas más específicas. Cada lenguaje tiene una categoría como "estructuras de datos" por ejemplo
